requirements.txt
python-irodsclient
apache-airflow
requests
